refactor: optimize medication table with reusable card components
